Might just fibonacci this

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 
+
 /**
 * main - entry
 * Return: 0 = success
@@ -7,19 +8,19 @@
 
 int main(void)
 {
-int val1 = 1;
+long int val1 = 1;
 
-int val2 = 2;
+long int val2 = 2;
 
 int count = 0;
 
-printf("%i, ", val1);
-printf("%i, ", val2);
+printf("%li, ", val1);
+printf("%li, ", val2);
 
 while (count < 50)
 	{
 	int lastvalue = val1 + val2;
-	
+
 	printf("%i, ", lastvalue);
 	val1 = val2;
 	val2 = lastvalue;
